Refactored into seperate files
While I'm following this tutorial the size of functions and files is starting to annoy me so im trying to break it up a little, not sure how to tackle the main js file "game" though.

I feel like it needs an entire rewrite, just following the tutorial to see what I can learn but the way the code is written is genuinely making me nauseous.